:recycle: Massive refactor is now mostly complete
:fire: Removed a BUNCH of files that are no longer needed
:heavy_plus_sign: Several dependencies have been added
:boom: This refactor resulted in everything changing. Now that we are
focusing on using poetry almost exclusively, a lot of the
distutils/setuptools related code is no longer needed. However, more
work is required to get everything up to date regarding the
configuration settings for CMake, as well as permitting the CMake code
to execute during a wheel install. Effectively, this library is unable
to execute properly for an install, and is only useful as a novelty when
importing. Work is being done to handle this, but it will be slow going
to due scheduling and other projects that require my attention.
Documentation of the cmake functions is also sorely needed.
Additionally, I am DESPERATELY in need of sitting down and getting the
tests to work in a way that I am satisfied. Currently, there is little I
can do to get pytest to work with this stuff and it's frankly going to
be a complete and total nightmare.

Continuous Integration (via Azure Pipelines) is also needed.